requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.2
matplotlib-inline==0.1.7
mysqlclient==2.2.4
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
requests==2.32.3
selenium==4.23.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.35.1
tornado==6.4.1
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
uvicorn==0.30.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
